+++ jwt-authorizer/src/lib.rs
@@ -6,9 +6,10 @@ use jsonwebtoken::TokenData;
 use serde::de::DeserializeOwned;
 
 pub use self::error::AuthError;
+pub use authorizer::{Authorizer, IntoLayer};
 pub use claims::{NumericDate, OneOrArray, RegisteredClaims};
 pub use jwks::key_store_manager::{Refresh, RefreshStrategy};
-pub use layer::{IntoLayer, JwtAuthorizer};
+pub use layer::JwtAuthorizer;
 pub use validation::Validation;
 
 pub mod authorizer;