+++ jwt-authorizer/tests/tests.rs
@@ -12,7 +12,12 @@ mod tests {
         BoxError, Router,
     };
     use http::{header, HeaderValue};
-    use jwt_authorizer::{layer::JwtSource, validation::Validation, IntoLayer, JwtAuthorizer, JwtClaims};
+    use jwt_authorizer::{
+        authorizer::Authorizer,
+        layer::{AsyncAuthorizationLayer, JwtSource},
+        validation::Validation,
+        IntoLayer, JwtAuthorizer, JwtClaims,
+    };
     use serde::Deserialize;
     use tower::{util::MapErrLayer, ServiceExt};
 
@@ -23,7 +28,7 @@ mod tests {
         sub: String,
     }
 
-    async fn app(jwt_auth: impl IntoLayer<User>) -> Router {
+    async fn app(layer: AsyncAuthorizationLayer<User>) -> Router {
         Router::new().route("/public", get(|| async { "hello" })).route(
             "/protected",
             get(|JwtClaims(user): JwtClaims<User>| async move { format!("hello: {}", user.sub) }).layer(
@@ -32,18 +37,22 @@ mod tests {
                         tower::buffer::BufferLayer::new(1),
                         MapErrLayer::new(|e: BoxError| -> Infallible { panic!("{}", e) }),
                     ),
-                    jwt_auth.into_layer().await.unwrap(),
+                    layer,
                 ),
             ),
         )
     }
 
-    async fn proteced_request_with_header(
-        jwt_auth: impl IntoLayer<User>,
+    async fn proteced_request_with_header(jwt_auth: JwtAuthorizer<User>, header_name: &str, header_value: &str) -> Response {
+        proteced_request_with_header_and_layer(jwt_auth.build().await.unwrap().into_layer(), header_name, header_value).await
+    }
+
+    async fn proteced_request_with_header_and_layer(
+        layer: AsyncAuthorizationLayer<User>,
         header_name: &str,
         header_value: &str,
     ) -> Response {
-        app(jwt_auth)
+        app(layer)
             .await
             .oneshot(
                 Request::builder()
@@ -56,15 +65,18 @@ mod tests {
             .unwrap()
     }
 
-    async fn make_proteced_request(jwt_auth: impl IntoLayer<User>, bearer: &str) -> Response {
+    async fn make_proteced_request(jwt_auth: JwtAuthorizer<User>, bearer: &str) -> Response {
         proteced_request_with_header(jwt_auth, "Authorization", &format!("Bearer {bearer}")).await
     }
 
     #[tokio::test]
     async fn protected_without_jwt() {
-        let jwt_auth: JwtAuthorizer<User> = JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem");
+        let auth: Authorizer<User> = JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem")
+            .build()
+            .await
+            .unwrap();
 
-        let response = app(jwt_auth)
+        let response = app(auth.into_layer())
             .await
             .oneshot(Request::builder().uri("/protected").body(Body::empty()).unwrap())
             .await
@@ -342,25 +354,46 @@ mod tests {
     // --------------------------
     #[tokio::test]
     async fn multiple_authorizers() {
-        let auths: Vec<JwtAuthorizer<User>> = vec![
-            JwtAuthorizer::from_ec_pem("../config/ecdsa-public1.pem"),
-            JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem").jwt_source(JwtSource::Cookie("ccc".to_owned())),
+        let auths: Vec<Authorizer<User>> = vec![
+            JwtAuthorizer::from_ec_pem("../config/ecdsa-public1.pem")
+                .build()
+                .await
+                .unwrap(),
+            JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem")
+                .jwt_source(JwtSource::Cookie("ccc".to_owned()))
+                .build()
+                .await
+                .unwrap(),
         ];
 
         // OK
-        let response =
-            proteced_request_with_header(auths, header::COOKIE.as_str(), &format!("ccc={}", common::JWT_RSA1_OK)).await;
+        let response = proteced_request_with_header_and_layer(
+            auths.into_layer(),
+            header::COOKIE.as_str(),
+            &format!("ccc={}", common::JWT_RSA1_OK),
+        )
+        .await;
         assert_eq!(response.status(), StatusCode::OK);
 
-        let auths: [JwtAuthorizer<User>; 2] = [
-            JwtAuthorizer::from_ec_pem("../config/ecdsa-public1.pem"),
-            JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem").jwt_source(JwtSource::Cookie("ccc".to_owned())),
+        let auths: [Authorizer<User>; 2] = [
+            JwtAuthorizer::from_ec_pem("../config/ecdsa-public1.pem")
+                .build()
+                .await
+                .unwrap(),
+            JwtAuthorizer::from_rsa_pem("../config/rsa-public1.pem")
+                .jwt_source(JwtSource::Cookie("ccc".to_owned()))
+                .build()
+                .await
+                .unwrap(),
         ];
 
         // Cookie missing
-        let response =
-            proteced_request_with_header(auths, header::COOKIE.as_str(), &format!("bad_cookie={}", common::JWT_EC2_OK))
-                .await;
+        let response = proteced_request_with_header_and_layer(
+            auths.into_layer(),
+            header::COOKIE.as_str(),
+            &format!("bad_cookie={}", common::JWT_EC2_OK),
+        )
+        .await;
         assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
         assert_eq!(response.headers().get(header::WWW_AUTHENTICATE).unwrap(), &"Bearer");
     }